refactor: rename navigation component to navbar
