+++ src/data/index.ts
@@ -1,3 +1,2 @@
 export * from './experiences'
-export * from './navItems'
 export * from './routes'
